refactoring, moving TermTree to its own file
